fix: now there is a validation if the user actually exists on session befoure logout

+++ src/controllers/users.controllers.ts
@@ -110,6 +110,10 @@ export async function logoutUser(req: Request, res: Response) {
         where: { token: userToken }
     })
 
+    if (!userData) {
+        return res.status(401).json({error: 'Token not found on sessions.'})
+    }
+
     await prisma.sessions.delete({
         where: { id: userData.id }
     });
